+++ lib/test-main.sh
@@ -2,7 +2,30 @@
 
 ble-import lib/core-test
 
-ble/test/start-section 'ble/main' 21
+ble/test/start-section 'ble/main' 29
+
+(
+  function f1 {
+    builtin eval -- "$_ble_bash_POSIXLY_CORRECT_local_adjust"
+    ble/util/setexit 123
+    builtin eval -- "$_ble_bash_POSIXLY_CORRECT_local_return"
+  }
+  function f2 {
+    builtin eval -- "$_ble_bash_POSIXLY_CORRECT_local_adjust"
+    [[ ! -o posix ]]
+    builtin eval -- "$_ble_bash_POSIXLY_CORRECT_local_return"
+  }
+  set +o posix
+  ble/test 'f1' exit=123
+  ble/test 'f2'
+  ble/test 'f1; [[ ! -o posix ]]'
+  ble/test 'f2; [[ ! -o posix ]]'
+
+  ble/test 'set -o posix; f1;                 ret=$?; set +o posix' ret=123
+  ble/test 'set -o posix; f2;                 ret=$?; set +o posix' ret=0
+  ble/test 'set -o posix; f1; [[ -o posix ]]; ret=$?; set +o posix' ret=0
+  ble/test 'set -o posix; f2; [[ -o posix ]]; ret=$?; set +o posix' ret=0
+)
 
 # ble/util/{put,print}
 (